Updating to complete version.

+++ webserver/browser_consumer.py
@@ -2,21 +2,14 @@ import simplejson, boto, uuid
 import os, thread, time
 
 ##
-## Functions to store and read messages with Amazon sqs.
+## Functions to retrieve and save updated pdfs from Amazon s3.
 ##
 
-local_file_store = '<local_file>' ## TODO: Update on server
 message_queue_name_request = 'webtex-tex-test-requests'
 message_queue_name_replies = 'webtex-tex-test-replies'
 
 message_bucket_name = 'webtext.text.com'
 
-do_update_message = {
-    "request": "update",
-    "file_path": "",
-    "file_name": "",
-    "bucket_name": "",
-}
 
 done_update_message = {
     "request": "done",
@@ -44,18 +37,24 @@ def read_message(message_queue_name):
   message = q.read()
   while message is None:   # if it is, continue reading until you get a message
     message = q.read()
+  print 'message is ', message.get_body()
   msg_data = simplejson.loads(message.get_body())
   key = boto.connect_s3().get_bucket(msg_data['bucket']).get_key(msg_data['key'])
-  data = simplejson.loads(key.get_contents_as_string())
-  q.delete_message(message)
-  return data
+  if key is not None:
+	  print 'key is ', key.get_contents_as_string()
+	  data = simplejson.loads(key.get_contents_as_string())
+	  q.delete_message(message)
+	  return data
+  else:
+	return None
 
 # Store update/done messages into the message queue
 def store_message(message_queue_name, message_bucket_name, message_bucket_key, message_structure):
   # sqs messages are contrained to 8kb/message. Store message data in a bucket on s3 - might be overkill!
   sqs = boto.connect_sqs()
   #this will return a new queue if one does not exist, or the existing queue
-  q = sqs.create_queue(message_queue_name) # Message queue points to message data.
+  #q = sqs.create_queue(message_queue_name) # Message queue points to message data.
+  q = sqs.get_queue(message_queue_name)
   data = simplejson.dumps(message_structure)
   s3 = boto.connect_s3()
   bucket = s3.get_bucket(message_bucket_name)
@@ -65,35 +64,42 @@ def store_message(message_queue_name, message_bucket_name, message_bucket_key, m
   q.write(message)
 
 ##
-## S9 File Storage Functions
+## S3 File Storage Functions
 ##
 
-def retrieve_file(bucket_name, file_key, filename):
+def retrieve_file(bucket_name, file_key, file_name):
   s3 = boto.connect_s3()
   key = s3.get_bucket(bucket_name).get_key(file_key)
-  key.get_contents_to_filename('/' + filename)
+  key.get_contents_to_filename(file_name)
 
-def store_file(bucket_name, file_key, filename):
-  s3 = boto.connect_s3()
-  bucket = s3.create_bucket(bucket_name)  # bucket names must be unique
-  key = bucket.new_key(file_key)
-  key.set_contents_from_filename(filename)
-  key.set_acl('public-read')
+# Parse request
+def do_some_work(data, file_name):
+  print 'working on data... ', data
+  if data['request'] == 'done':
+    print 'performing update...'
+    print data
+    bucket_name = 'test_bucket.webtex.com'
+    file_path = 'LATEX0/'
+    file_name = 'temp.zip'
+    #retrieve_file(data['bucket_name'], data['file_path'] + data['file_name'], file)
+    retrieve_file(bucket_name, file_path + file_name, file_name)
+    print 'file should have been retrieved, yaldi'
+    return True
+  
+  print 'request != update'
+  return False
+
+def main(file_name):
+  data = read_message(message_queue_name_replies)
+  if data is not None:
+    print 'got data ', data
+    #thread.start_new_thread(do_some_work, (data, ))
+    return do_some_work(data, file_name)
+
+  print 'naw'
+  return False
+
 
 # Main daemon 
-if __name__ == '__main__':
-	create_queue(message_queue_name_request, 120)
-	create_bucket(message_bucket_name)
-	count=0
-	print 'queues are created'
-	# TODO Update on server to update zip file to be processed periodically
-	while(True):
-		do_update_message['file_path'] = 'LATEX%i/' % (count)
-		count += 1
-		do_update_message['file_name'] = 'paper.tar.gz'
-		do_update_message['bucket_name'] = 'test_bucket.webtex.com'
-		print 'Updating ' + do_update_message['file_path'] + do_update_message['file_name']
-		store_file(do_update_message['bucket_name'], do_update_message['file_path'] + do_update_message['file_name'], local_file)
-		store_message(message_queue_name_request, message_bucket_name, 'text_key', do_update_message)
-		print 'file stored, message update in queue'
-		time.sleep(10) # Timeout in sec
+#if __name__ == '__main__':
+#  main()
